// frontend/src/interfaces/Participant.tsx
import Perk from "./Perk";

interface Participant {
    puuid: string;
    teamId: number;
    spell1Id: number;
    spell2Id: number;
    championId: number;
    profileIconId: number;
    riotId: string;
    bot: boolean;
    summonerId: string;
    gameCustomizationObjects: any[];
    perks: Perk;
    predictedRole: string;
}

export default Participant